Re render and fix closure
